Added check for zero valid sites Upgraded doxygen and fixed doxygen comments Fixed bug in write results Changed computation order. First H1 then H0 Changed option init H1 from H0 to init H0 from H1 Added maximization forced stop if result does not already pass the LRT Added MPI trace level and fixed messages Added testing hooks to BEB. Passed scale values from H1 to BEB Fixed relative error default Added initial map of internal to external site numbers Changed iterator initialization
git-svn-id: https://svn.vital-it.ch/svn/hp2c/trunk/Codeml_Variants/Fastcodeml@4924 95c5a404-1f88-0410-a6b0-c3f062b6f34a

+++ CodonFrequencies.h
@@ -58,7 +58,7 @@ public:
 
 		// Support values needed for the eigensolver
 		mNumGoodCodons = 0;
-		for(size_t k=0; k < N; ++k)
+		for(size_t k=0; k < static_cast<size_t>(N); ++k)
 		{
 			mCodonFreqSqrt[k] = sqrt(mCodonFrequencies[k]);
